changed parseFloat to parseInt since that would be more appropriate

+++ api/canvas.js
@@ -4,7 +4,7 @@ var db = require('../database.js');
 
 module.exports = {
     list: function(req, resp) {
-        let limit = (req.query.limit === undefined) ? NaN : parseFloat(req.query.limit);
+        let limit = (req.query.limit === undefined) ? NaN : parseInt(req.query.limit);
         db.query("SELECT canvas_id AS 'id', disp_name AS 'name', width, height, c_date AS 'date' FROM " + config.database.prefix + "canvas_list ORDER BY c_date DESC"
                 + ((!isNaN(limit)) ? (" LIMIT " + limit) : ""), function(err, result, fields) {
             if(err) resp.status(500).send(template(null, err + ''));
@@ -13,7 +13,7 @@ module.exports = {
     },
 
     fetch: function(req, resp) {
-        let id = parseFloat(req.params.id);
+        let id = parseInt(req.params.id);
         if(isNaN(id)) resp.status(400).send(template(null, 'Invalid canvas ID'));
         db.query("SELECT tab_name FROM " + config.database.prefix + "canvas_list WHERE canvas_id = " + id, function(id_err, id_result, id_fields) {
             if(id_err) resp.status(500).send(template(null, id_err + ''));
@@ -31,8 +31,8 @@ module.exports = {
     },
 
     history: function(req, resp) {
-        let id = parseFloat(req.params.id);
-        let offset = parseFloat(req.params.offset);
+        let id = parseInt(req.params.id);
+        let offset = parseInt(req.params.offset);
         if(isNaN(id)) resp.status(400).send(template(null, 'Invalid canvas ID'));
         if(isNaN(offset)) resp.status(400).send(template(null, 'Invalid offset'));
         db.query("SELECT tab_name, width, height FROM " + config.database.prefix + "canvas_list WHERE canvas_id = " + id, function(id_err, id_result, id_fields) {
